extract-clients: fix imprecise timestamps at beginning of chunks
Previously packets in all chunks had timestamp 0 and ignored original
time difference from the last packet in the previous chunk.

For steady traffic with non-negligible QPS it does not make significant
difference, but it leads to big imprecision if original traffic had
"gaps" in it.

After this change first chunk starts at time 0 and timing of all packets
is relative to timing of the very first packet in first chunk. Empty
chunks with no packets will be created if gap in traffic is long enough.