merge_chunks: utility to merge chunks on-the-fly
Intended usage is to avoid generating PCAPs which are simple combination
of "base chunks".

Example:
When original PCAP is split into 100 chunks with 1 kQPS on average,
testing full range of 1k to 100k QPS formerly required generating 100
distincts PCAPs, wasting time and storage.
With this utility it is enough to generate "base chunks" and create
arbitrary multiplication on-the-fly.

Why Python?
- Lua cannot list content of directory, FFI solution would be
  unportable, and adding depedency just for that seems too much.
- BASH version of this script made me cry when I finished it.
- Python is already depedency of replay.py and it is used only to
  process mergecap arguments and python process terminates/replaces
  itself once its job is done.