filter-dnsq: skip 'special' queries for *.dotnxdomain.net
By default, filter out queries for subdomains of dotnxdomain.net.
This is a 'special' measurement domain. Queries directed to it have
timestamps encoded in qname and replaying old queries results in
timeouts, not in a realistic traffic replay.

A new option -s can be used to keep the queries in the output if
desired.

The other domain - dashnxdomain.net - did not appear in any of my PCAPs
so for simplicity I skipped it.

Implementing this check forced me to parse more parts of DNS header, but
even for tens of GBs of data it was so fast that I do not consider it to
be a problem.

Fixes: #25