filter-dnsq: always parse packets
Formerly, if malformed packets were requested in the output,
the script skipped DNS parsing because there was no point in doing so:
The malformed packets would have been included in the output anyway.

As preparation for a new feature which requires access to qname,
the script now tries to parse the packet even if user requested
inclusion of malformed packets in the output.

Even for tens of GBs of data overhead of doing this was negligible
compared to other processing, so I did not go to the trouble of
optimizing this further.