Change linestyle in latency charts
When plotting multiple different versions together, it is useful if they
can be visually told apart by using a different linestyle. Since the
only input data is the file path and the shotgun JSON itself, provide
the user with an option to change the linestyle based on the file name,
which is also used as the chart label.

A choice to use regular expression was made to keep maximum flexibility
for future use. The supported linestyles are limited to four predefined
named choices which matplotlib supports [1].

[1] - https://matplotlib.org/stable/gallery/lines_bars_and_markers/linestyles.html