plot-packet-rate: handle incomplete last sampling period
More often than not, the last sampling period in CSV is incomplete and
that used to cause the plotter to not print sampling period in the
chart, rendering the field useless.

Ignore the change in the sampling period if it happened on the very last
sample and use preceding (constant) period value in generated chart.