Enhanced default TLS priority
This allows us to use a `dnssim-default` keyword at the beginning of a
GnuTLS priority string, which lets dnssim choose the proper priority for
different protocols. As QUIC disallows some of the less secure options,
we can change the priority accordingly but still allow the user to
modify our own defaults like this.