Use CMake to build dnssim
Replaces GNU Autotools with CMake, which seems to be taking over the
world of building C projects.

As for dnssim, this makes it (in my opinion) slightly easier to
maintain, as Autotools tend to be quite clunky when settings change,
which is quite often the case when introducing new libraries (like
ngtcp2 for QUIC - !75).

It will also make it kind of easier to use Git submodules for some
dependencies (again handy for ngtcp2, which is not yet stable).