Generic multi-stream connections (QUIC preparation)
This commit makes it so that all connections - not just HTTPS/2 - may
support multiple streams at once. Of course, still only HTTPS/2 actually
uses this, other connection types only use a '-1' dummy stream ID
because the protocols themselves do not support multiple streams, but
this prepares the system for QUIC and pre-emptively de-duplicates the
required logic.