Assert dnssim load
When containerized, Shotgun would randomly fail with `Undefined symbol:
output_dnssim_new`. When I add this assertion, it somehow seems to work
reliably, so I assume that LuaJIT might be throwing away the seemingly
unused result of `loader.load` sometimes, which causes the issue.

Another solution *might* be just using the returned namespace directly,
but I don't really want to go rewriting the whole `.lua` file for this,
if the assert works.