Abort traffic replay if servers are unresponsive
If the target server doesn't send any reply within the configured time
limit (when some are expected), abort the traffic replay. This most
commonly happens due to either incorrect configuration or server crash.
In the case of premature exit, ensure that existing statistics are
collected.

The default (baked-in) server timeout limit is configured to be 16
seconds. This has proven to be the right magic number in the development
of related tools (respdiff). It is unlikely that this limit would have
to be reconfigured or disabled, so no way to adjust or turn off this
limit is provided until a reasonable use-case is presented.