Abort if the number of discarded packets exceeds limit
Discarded packets indicate either invalid packets in the input, or a
runtime issue during replay. To avoid processing tainted data, abort the
traffic replay with an error if that happens.

The threshold of discarded packets is set at 0.1 % of the total packets.
Each traffic sender thread does its own accounting of this limit. The
threshold is only checked once 100 discarded packets are detected to
provide some leeway when the packet numbers are low.